Rename ShowAllVariableXxx to ShowAllVariablesXxx.
