+++ shardingsphere-test/shardingsphere-parser-test/src/main/resources/case/ral/queryable.xml
@@ -18,7 +18,7 @@
 
 <sql-parser-test-cases>
     <show-variable sql-case-id="show-variable" name="CACHED_CONNECTIONS" />
-    <show-all-variable sql-case-id="show-all-variables" />
+    <show-all-variables sql-case-id="show-all-variables" />
     
     <show-readwrite-splitting-read-resources sql-case-id="show-readwrite-splitting-read-resources" />
